+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.1.4 (2019-09-16)
+------------------
+Reads hydro h5 flows and area
+Godin filter time series
+command line tool :
+ extract_dss to postprocess to daily, max, min and monthly from DSM2 dss files
+
 0.1.0 (2019-08-21)
 ------------------